Settings refactoring
