+++ src/components/recentPosts.ts
@@ -1,7 +1,8 @@
 import { SettingsFormField, SettingsValues, TriggerContext } from "@devvit/public-api";
-import { GeneralSetting, IncludeRecentContentOption, selectFieldHasOptionChosen } from "../settings.js";
+import { GeneralSetting } from "../settings.js";
 import { getSubredditName } from "../utility.js";
 import markdownEscape from "markdown-escape";
+import { IncludeRecentContentOption, selectFieldHasOptionChosen } from "../settingsHelpers.js";
 
 enum RecentPostsSetting {
     IncludeRecentPosts = "includeRecentPosts",
